+++ BanchoNET.Core/Models/Players/ModeStats.cs
@@ -6,7 +6,7 @@ public class ModeStats
 {
 	public long TotalScore { get; set; }
 	public long RankedScore { get; set; }
-	public ushort PP { get; set; }
+	public int PP { get; set; }
 	public float Accuracy { get; set; }
 	public int PlayCount { get; set; }
 	public int PlayTime { get; set; }
